+++ controller/utilscontroller.php
@@ -87,7 +87,17 @@ class UtilsController extends Controller {
     public function __construct($AppName, IRequest $request, $UserId,
         $userfolder, $config, IAppManager $appManager){
         parent::__construct($AppName, $request);
-        $this->appPath = $appManager->getAppPath('gpxpod');
+        // just to keep Owncloud compatibility
+        // the first case : Nextcloud
+        // else : Owncloud
+        if (method_exists($appManager, 'getAppPath')){
+            $this->appPath = $appManager->getAppPath('gpxpod');
+        }
+        else {
+            $this->appPath = \OC_App::getAppPath('gpxpod');
+            // even dirtier
+            //$this->appPath = getcwd().'/apps/gpxpod';
+        }
         $this->userId = $UserId;
         $this->dbtype = $config->getSystemValue('dbtype');
         if ($this->dbtype === 'pgsql'){